fix(slicer-sdk): cast layer_index to i32 in support_plan_segments_for filter
PaintRegionLayerView::layer_index is u32 but SupportPlanEntry.global_layer_index
is i32. The filter now compares `self.layer_index as i32 == entry.global_layer_index`.

+++ crates/slicer-sdk/src/traits.rs
@@ -129,7 +129,7 @@ impl PaintRegionLayerView {
         plan.entries
             .iter()
             .filter(|entry| {
-                entry.global_layer_index == self.layer_index
+                entry.global_layer_index == self.layer_index as i32
                     && entry.object_id == object_id
                     && entry.region_id == region_id
             })
